backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-finance-analyzer"
version = "0.1.0"
description = "AI Personal Finance Analyzer — FastAPI Backend"
requires-python = ">=3.12"
dependencies = [
    # ── Web Framework ──────────────────────────────────────────
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",

    # ── Database ───────────────────────────────────────────────
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",

    # ── Validation & Config ────────────────────────────────────
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "pydantic[email]>=2.7.0",
    "python-dotenv>=1.0.0",

    # ── Auth ───────────────────────────────────────────────────
    "python-jose[cryptography]>=3.3.0",
    "argon2-cffi>=23.1.0",

    # ── AWS ────────────────────────────────────────────────────
    "boto3>=1.34.0",

    # ── Kafka ──────────────────────────────────────────────────
    "aiokafka>=0.11.0",

    # ── AI / ML ────────────────────────────────────────────────
    "openai>=1.35.0",
    "scikit-learn>=1.5.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",

    # ── File Parsing ───────────────────────────────────────────
    "pdfplumber>=0.11.0",

    # ── Payments ───────────────────────────────────────────────
    "stripe>=10.0.0",

    # ── Email ──────────────────────────────────────────────────
    "jinja2>=3.1.0",

    # ── HTTP Client ────────────────────────────────────────────
    "httpx>=0.27.0",

    # ── Logging ────────────────────────────────────────────────
    "structlog>=24.0.0",

    # ── Resilience ─────────────────────────────────────────────
    "tenacity>=8.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "respx>=0.21.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
    "types-stripe",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

# ── Ruff (linter + formatter) ──────────────────────────────────
[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "ANN"]
ignore = ["ANN101", "ANN102", "ANN401"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ANN"]
"alembic/*" = ["ANN", "F401"]

# ── Mypy ──────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
exclude = ["alembic/"]

# ── Pytest ────────────────────────────────────────────────────
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=app --cov-report=term-missing --cov-report=xml"
